requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=h27087fc_1
absl-py=1.4.0=pyhd8ed1ab_0
aiohttp=3.8.1=py39hb9d737c_1
aiosignal=1.3.1=pyhd8ed1ab_0
appdirs=1.4.4=pyhd3eb1b0_0
async-timeout=4.0.2=py39h06a4308_0
attrs=23.1.0=pyh71513ae_1
blas=1.0=mkl
blinker=1.6.2=pyhd8ed1ab_0
bottleneck=1.3.5=py39h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py39hb9d737c_1004
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pyhd8ed1ab_0
certifi=2023.5.7=py39h06a4308_0
cffi=1.15.0=py39h4bc2ebd_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=unix_pyhd8ed1ab_2
contourpy=1.0.5=py39hdb19cb5_0
cryptography=3.4.8=py39h95dcef6_1
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py39h5eee18b_0
geomdl=5.3.1=py_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.18.0=pyh1a96a4e_0
google-auth-oauthlib=1.0.0=pyhd8ed1ab_0
grpc-cpp=1.48.2=h5bf31a4_0
grpcio=1.48.2=py39h5bf31a4_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=58.2=he6710b0_3
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.6.0=pyha770c72_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2023.1.0=hdb19cb5_46305
joblib=1.1.1=py39h06a4308_0
jpeg=9b=h024ee3a_2
kiwisolver=1.4.4=py39h6a678d5_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libuv=1.44.2=h5eee18b_0
libwebp=1.2.0=h89dd481_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.3=pyhd8ed1ab_0
matplotlib=3.6.2=py39h06a4308_0
matplotlib-base=3.6.2=py39h945d387_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.6=py39h417a72b_1
mkl_random=1.2.2=py39h417a72b_1
multidict=6.0.2=py39h5eee18b_0
munkres=1.1.4=py_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numexpr=2.8.4=py39hc78ab66_1
numpy=1.19.5=py39h7820934_5
numpy-base=1.19.5=py39hfdd66db_5
oauthlib=3.2.2=pyhd8ed1ab_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
packaging=23.0=py39h06a4308_0
pandas=1.3.5=py39h8c16a72_0
pcre=8.45=h295c915_0
pillow=9.3.0=py39hace64e9_1
pip=23.0.1=py39h06a4308_0
pooch=1.4.0=pyhd3eb1b0_0
protobuf=3.20.3=py39h6a678d5_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.7.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=3.0.9=py39h06a4308_0
pyqt=5.9.2=py39h2531618_6
pysocks=1.7.1=pyha2e5f31_6
python=3.9.16=h7a1cb2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.9.1=py3.9_cuda10.2_cudnn7.6.5_0
pytz=2022.7=py39h06a4308_0
pyu2f=0.1.5=pyhd8ed1ab_0
qt=5.9.7=h5867ecd_1
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
requests=2.29.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.2.0=py39h6a678d5_0
scipy=1.10.1=py39hf6e8229_1
setuptools=58.0.4=py39h06a4308_0
sip=4.19.13=py39h295c915_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.13.0=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py39h52d8a92_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.9.1=py39
torchvision=0.10.1=py39_cu102
tornado=6.2=py39h5eee18b_0
typing_extensions=4.5.0=py39h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.15=pyhd8ed1ab_0
werkzeug=2.1.2=pyhd8ed1ab_1
wheel=0.38.4=py39h06a4308_0
xz=5.4.2=h5eee18b_0
yarl=1.7.2=py39hb9d737c_2
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.4.9=haebb681_0
